requirements.txt
openai
PyYAML
